requirements.txt
streamlit==1.32.0
python-dotenv==1.0.0
scikit-learn==1.4.0
xgboost==2.0.3
pandas==2.1.0
numpy==1.26.0
matplotlib==3.8.0
plotly==5.18.0
requests==2.31.0